Add TestObject and refactor Object

+++ src/game/types/Types.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-using Float = double;
+using Float = float;
